migrate to spring boot

+++ core/core-server/src/main/java/by/wadikk/core/model/Article.java
@@ -4,12 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Entity class Article
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Article {
 
-    private Long id;
+    private Integer id;
     private Product product;
     private Store store;
     private double price;
